feat: add DQuickBlitFramebuffer
In the opengl backend, the DQuickBlitFramebuffer used by the
QOpenGLFramebufferObject::blitFramebuffer to copy the current frame
color buffer. In the QPainter backend, the DQuickBlitFramebuffer
to copy to a QImage from the paint device and the device is a QImage
or a QPixmap. In the other QQuickItem, we can use the
DQuickBlitFramebuffer to capture a QSGTexture, next, it can be made to
do some processing, for example, use it with QtGraphicalEffects can
do Gaussian blur in a region in real time.

Log:
Change-Id: Ia3936518ab5d1eb7ad82603df67b71c6701666eb