fix: Remove id refrences between sibling elements
When you use a control and inherit one of the elements, if other
elements refer to the element's id, an error message will appear
that the element is not present. Only by not referring to the ID
of the sibling element can we safely construct the style of a
control.

Log:
Influence: Dtk qml

Change-Id: Ifb37fcad43238c14eed941c4b699bc4851376ca3