feat: Add AlertToolTip control and adjust LineEdit control
Like dtkwidget, the AlertToolTip control will show alert message if the LineEdit
applied a error or warning. Use it within qml to add some warning messages
is more easier than dtkwidget, for example:

```
LineEdit {
    alertText: "This is a long sentence."
    alertDuration: 3000
    showAlert: focus
}

or use it in a custom item:
Item {
    AlertToolTip {
        target: parent
        alertText: "This is a long sentence."
    }
}
```
It has the same function as the control in dtkwidget, but some customized
functionality will be added in subsequent updates.

Log:
Influence: LineEdit qml
Change-Id: Ib55b236df04cc816cdb4211a44fa5fb04e281e03