refactor: Adjust SearchEdit control and add the animation for it
The old SearchEdit control is disorderly in layout and strange in style. Readjust
its layout, use rawlayout instead of absolute coordinates, and add animation when
the state changes to ensure fluency.

A shadow needs to be added to the rectange border in the new design. Because the
shadow of DTK has a small problem at present, it cannot be realized.

Log:
Influence: qml search edit and rectange border
Change-Id: Iba5f0fda3b2fa437bb9b41016bbb4f445a75a936