fix: The Application may crash on exit
1. After exiting, the program will update the enable status of
QML item, and then update the icon resource file. However, at
this time, the QML engine has been released, so the correspon-
ding ImageProvider cannot be read, resulting in the access to
the wild pointer crash.  Determine if the Engine still exists
while processing the Enable state to resolve this issue.

2. Some controls have the disable state on their first displa-
y after creation, because the diable bit in the color selecto-
r is not initialized and a random value may appear.

3. Fixed compilation errors and run failures under the QT-5.11
environment.

Log:
Influence: None
Change-Id: I526c49d20817e80796cae5d6ced78362a0276599