fix: Modify the SpinBox control style
1. Refactor the SpinBox control which provides only one style for a up
and down indicator arrow, and replaces the Qt native SpinBox
control style with it. The background of the indicator is independent
item, not the part of the icon, because sometimes there is only one
indicator present

2. Add a SpinBox of plus and minus button type. This class has only the
style of plus and minus arrows. Add a reset button for needing to reset
the data.

Log:
Influence: None
Change-Id: I6ed35b11b25e652f73a723364db1f6323dc6590a