feat: add BoxShadow control
use the shadowcolor to draw the source control, do blur
processing on its content, and shadow the blur result.
The inner shadow algorithm is different from the outer shadow algorithm.
it is cut on the outer shadow algorithm, and only the content located inside
the source control is retained.
if necessary, enable the cache option to cache shadow pictures,
please note that this option will increase memory consumption,
please use it reasonably

Log:
Change-Id: I9fb5e86c8c71476932c25fdc45fe962e0b60f484