fix: Hollow BoxShadow dont follow the geometry animation
The update of hollow painting needs to follow the update of the source
texture. When the texture is updated, the old node needs to be released.
Otherwise, the old node still needs to be used instead of releasing the
old texture every time, so as to solve the problem that the control
cannot display normally in the hollow state.

Log:
Influence: DTK BoxShadow
Change-Id: I3418ccb2d8304e7fbe4cfc74c30c10581fe0214c