feat: add the "Button" examples to exhibition demo
Supported to view the QML sources of examples.
Supported to preview for edit the examples's code.
Supported set the control state for the Button.
The ColorSelector support to follow it's parent "hovered" "pressed"
"disabled" "inactived" properites.
Fix warnings of HighlightPanel if the object of it "control" property
is not the Control type.

Log:
Change-Id: Ia6785d97869b959e3adb6c2e031b5bcab3a8db5d