fix: customize SpinBox indicator error
this is maybe a Qt bug, when setting value and customized indicator
for SpinBox, indicator is invalid to press, but it's ok if we add a indicator
binding when evaluate implicitWidth.
it should be related with time for setValue and indicator'initialization.

expose function to get control's background implicitWidth and
content implicitWidth.

Log: 
Influence: none
Change-Id: Id4b57c41d0665e4d7cd84cc29921660eaea492d3