fix: use a static image replace invalid image for icon
If it's can't get a valid QImage for the icon, will return a null
QImage value, it's get a warning message in Qt's QQuickImage. But
the warning message for a invalid icon is not friendly, so return
a 10x10 full black color QImage for the invalid icon on release.

Log:
Change-Id: Ib685e42078ff261b83af14d2d53b8766298c7ddb