fix: don't set the font property for the Control type items
The Control type items is inherit the font property from it's parent,
for a example:

Control {
    id: control2
    Control {
        id: control3
        // The font's resolve mask of control4 should be is 0, the
        // controls3's font is same thing. because control3's font
        // resolve mask is 0 also, so the font's all value equal to
        // the parent font values.
        font: control4.font
    }
}
Control {
    id: control0
    font.pointSize: 10
    Control {
        id: control4
    }
}

The font's resolve mask of control4 should be is 0, the controls3's font is
same thing. because control3's font resolve mask is 0 also, so the font's
all value equal to the parent's font values. And if exec the
"control0.font.pixelSize = 14", can't to change the control3's
font.pixelSize property, it's value keep equal to the control2's font.

Log:
Change-Id: I0a8ba24de1932c6c37a05a49f1df8a08c74e6be1