feat: Add implementation to accelerate process startup
Implements a way to speed up DTK program startup by using Preload,
which separates project source code through two different plugins,
one for the Preload plug-in displayed in the main window and the
other for the MainComponent plug-in that creates all components of
the program.

This method can increase the program start time to a critical point
and delay all time-consuming operations until the program starts.
Therefore, it can greatly guarantee the starting speed of the program.

The Preload plugin enables smooth design through animation, as well as
simulating false start of a program by overwriting a static graph.

Change-Id: Ic47566e3dc999b10e0ecdd81734e72b44b4e5781