fix: DQuickControlColorSelector's properties is undefined
The QtQml runtime will create a QQmlPropertyCache for the QObject in qml
context on first get property of the QObject in qml context, and store
the cache object to QMap<QObject*, QQmlPropertyCache*>. The
DQuickControlCorlorSelector's QMetaObject is dynamic meta object, it's
using QQmlOpenMetaObject to create properties on runtime, maybe some
properties is create after QQmlPropertyCache create, though
QQmlOpenMetaObject::createProperty will release the QQmlPropertyCache
for the QObject, but the QML engine will get the cache from
QMap<QObject*, QQmlPropertyCache*> on later. So maybe some
DQuickControlColorSelector's dynamic properties is not in
QQmlPropertyCache, give rise to "undefined" when using those properties
in qml.

Bug: #117