fix: Missing Qml registration info in test project
  Refactor the CMake code for better code reusability between qt5
and qt6.
  the testing project is no longer compiled together with the
source code for code coverage, we add a option `ENABLE_COV`, and
recompile the project.
Because the tests also load QML plugins that depend on libraries
compiled from the source code, it leads to conflicts between the
binary-loaded libraries and the original ones.
As a result, the registration information within QML is lost
when loading QML plugins.
  Add EventLoop for unit-test.
  UT is wrong for WindowButtonGroup in qt6.