fix: crash at a nullptr QQuickWindow pointer
  It sames as the commit 4cf8a6b7ec3ca944d21e4c88831581adb38a0a00
We store window variable of the item to avoid invaild in some step,
because render function is working in readerer thread and `window`
of the item is in main thread, it maybe is nullptr in sometime,
e.g: QQuickItem::setParentItem in QQuickPopup when switching visible.
  We reserve the judgment of `m_item->window()` to reduce the possibility
of errors in render function.
  TODO: QSGNode's render function shouldn't access QQuickItem.

Issue: https://github.com/linuxdeepin/developer-center/issues/5228