fix: app crashed when when InWindowBlur switching frequently
updatePaintNode will be executed  multiple times, so it should at least
be a slot function that only connects once, and `node` maybe
a invalid pointer, because QSGNode and QuickItem are managed by
two different threads.
We can use `m_item` to adjust whether the texture is valid, because
it's a QPointer for QQuickItem.